requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pillow==10.4.0
python-amazon-paapi==5.0.1
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==70.3.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.2
